feat: 重构 HistogramChart 组件，拆分为多个模块并添加数据处理和事件处理功能
