+++ miaocha-ui/frontend/src/pages/Home/HistogramChart/export.ts
@@ -1,0 +1,7 @@
+// 导出所有模块
+export { default as HistogramChart } from './index';
+export * from './types';
+export * from './utils';
+export * from './chartConfig';
+export * from './eventHandlers';
+export * from './dataProcessor';
